requirements.txt
# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Job Queue
redis==5.0.1
rq==1.16.0
rq-dashboard==0.8.6

# HTTP Clients
httpx>=0.24.0,<0.25.0
requests==2.31.0

# AI Providers
openai==2.14.0
google-generativeai==0.8.0
openai-agents==0.6.4

# Data
pydantic==2.5.3
python-dotenv==1.0.0
supabase==2.3.0

# Google Sheets
gspread==6.0.0
google-auth==2.27.0

# Scraping
firecrawl-py==0.0.16
apify-client>=1.5.0
beautifulsoup4==4.12.3
lxml==5.1.0
html5lib==1.1
parsel==1.8.1
cssselect==1.2.0
fake-useragent==1.4.0

# Scraping - Browser automation
playwright==1.41.0
selenium==4.17.0

# Scraping - Async & Rate limiting
aiohttp==3.9.3
tenacity==8.2.3
ratelimit==2.2.1

# Data Processing
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1

# Utilities
python-dateutil==2.8.2
